Fix AAC Huffman symbol base indexing, SBR QMF windowing, and container memory leaks
- Fix Huffman quad decoding base dimensions for books 3 and 4 in libfaad/huffman.c
- Fix Huffman pair decoding base dimensions for books 7 and 8 in libfaad/huffman.c
- Implement 640-tap QMF prototype synthesis filter windowing in libfaad/sbr.c
- Filter MP4 box parsing to only process audio tracks and clean up track sample table allocations in frontend/mp4read.c
- Implement in-memory PCMFifo sample ring buffer for gapless trimming in frontend/faad_main.c
- Fix 64-bit MP4 sample offsets in frontend/mp4read.c and frontend/faad_main.c

+++ frontend/faad_main.c
@@ -270,9 +270,9 @@ int main(int argc, char **argv)
             return 1;
         }
         for (uint32_t s = 0; s < track.num_samples; s++) {
-            uint32_t offset = track.samples[s].offset;
+            uint64_t offset = track.samples[s].offset;
             uint32_t size = track.samples[s].size;
-            if (offset > 0 && offset + size <= (uint32_t)file_len) {
+            if (offset > 0 && offset + size <= (uint64_t)file_len) {
                 uint8_t adts_hdr[7] = { 0xFF, 0xF1, 0x50, 0x80, 0x00, 0x1F, 0xFC };
                 uint32_t frame_len = size + 7;
                 adts_hdr[3] = (uint8_t)(0x80 | ((frame_len >> 11) & 0x03));
@@ -351,9 +351,9 @@ int main(int argc, char **argv)
 
     if (is_mp4) {
         for (uint32_t s = start_frame; s < track.num_samples; s++) {
-            uint32_t offset = track.samples[s].offset;
+            uint64_t offset = track.samples[s].offset;
             uint32_t size = track.samples[s].size;
-            if (offset == 0 || offset + size > (uint32_t)file_len) continue;
+            if (offset == 0 || offset + size > (uint64_t)file_len) continue;
 
             uint32_t bytes_consumed = 0;
             uint32_t bytes_written = 0;
